RYA-377 Repackaged the common Aggregation code into the rya.api.functions project.
